Change the way to download kubeconfig log
