+++ libs/ui-lib/lib/ocm/components/clusterDetail/utils.tsx
@@ -3,7 +3,7 @@ import { isInOcm, handleApiError, getApiErrorMessage } from '../../../common/api
 import { AlertsContextType, hostStatusOrder } from '../../../common';
 import { ClustersAPI } from '../../services/apis';
 import { ClustersService } from '../../services';
-import { stringToJSON } from '../../../common/utils';
+import { downloadFile, stringToJSON } from '../../../common/utils';
 import {
   Cluster,
   Host,
@@ -23,10 +23,10 @@ export const downloadClusterInstallationLogs = async (
         hostId: undefined,
         logsType: 'all',
       });
-      download(data.url);
+      downloadFile(data.url);
     } else {
       const { data, fileName } = await ClustersService.downloadLogs(clusterId);
-      download('', data, fileName);
+      downloadFile('', data, fileName);
     }
   } catch (e) {
     handleApiError(e, (e) => {
@@ -96,22 +96,4 @@ export const getMostSevereHostStatus = (hosts: Host[]) => {
     }
   }
   return status;
-};
-
-const download = (fileUrl?: string, dataBlob?: Blob, fileName?: string) => {
-  const link = document.createElement('a');
-  if (fileUrl && fileUrl !== '') {
-    link.setAttribute('href', fileUrl);
-  }
-  if (dataBlob) {
-    const file = new Blob([dataBlob], { type: 'text' });
-    link.setAttribute('href', URL.createObjectURL(file));
-  }
-  if (fileName) {
-    link.setAttribute('download', fileName);
-  }
-  link.style.visibility = 'hidden';
-  document.body.appendChild(link);
-  link.click();
-  document.body.removeChild(link);
 };
